+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 ### Changed
+- Use MIPMAP for heatmap layer.
 
 ## [1.1.9](https://www.npmjs.com/package/vitessce/v/1.1.9) - 2021-05-07